// rotv-flight/src/gen/rotv/v1/common_pb.d.ts
export enum SystemStatus {
  /**
   * @generated from enum value: SYSTEM_STATUS_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: SYSTEM_STATUS_NOMINAL = 1;
   */
  NOMINAL = 1,

  /**
   * @generated from enum value: SYSTEM_STATUS_WARNING = 2;
   */
  WARNING = 2,

  /**
   * @generated from enum value: SYSTEM_STATUS_CRITICAL = 3;
   */
  CRITICAL = 3,

  /**
   * @generated from enum value: SYSTEM_STATUS_OFFLINE = 4;
   */
  OFFLINE = 4,

  /**
   * @generated from enum value: SYSTEM_STATUS_UNKNOWN = 5;
   */
  UNKNOWN = 5,
}

/**
 * Describes the enum rotv.v1.SystemStatus.
 */
export declare const SystemStatusSchema: GenEnum<SystemStatus>;

/**
 * @generated from enum rotv.v1.Severity
 */
export enum Severity {
  /**
   * @generated from enum value: SEVERITY_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: SEVERITY_INFO = 1;
   */
  INFO = 1,

  /**
   * @generated from enum value: SEVERITY_OK = 2;
   */
  OK = 2,

  /**
   * @generated from enum value: SEVERITY_WARNING = 3;
   */
  WARNING = 3,

  /**
   * @generated from enum value: SEVERITY_CRITICAL = 4;
   */
  CRITICAL = 4,
}